Move view helper to paginator component
